Gradual transition to st instead of urxvt.
My st build works by leveraging tmux. The basic tmux.conf I have set up is meant to provide the extra functionality that st lacks out of the box. The idea is to use st without having to patch it heavily. As such, my tmux is not used to its potential as a multiplexer.